Suppressing useless files
