--- Grid.h
@@ -1,19 +1,0 @@
-#pragma once
-#include <vector>
-#include <iostream>
-#include "Token.h"
-#include <random>
-
-
-class Grid
-{
-private : 
-    vector<vector<Token*>> grid;
-public : 
-    //Constructors
-    Grid();
-    //Methods
-    void initGrid();
-    void updateGrid();
-
-}
