--- headers/Door.h
@@ -1,31 +1,0 @@
-// Class for the behavior of doors
-
-#pragma once
-#include "SpaceObject.h"
-
-class Door : public SpaceObject {
-
-public:
-	Door() : SpaceObject() {
-		idDoor = nbDoors;
-		nbDoors++;
-	}
-	Door(int x, int y) : SpaceObject(x, y) {
-		idDoor = nbDoors;
-		nbDoors++;
-	}
-	Door(bool red, bool white, bool blue) : SpaceObject(red, white, blue) {
-		idDoor = nbDoors;
-		nbDoors++;
-	}
-	Door(int x, int y, bool red, bool white, bool blue) : SpaceObject(x, y, red, white, blue) {
-		idDoor = nbDoors;
-		nbDoors++;
-	}
-
-private:
-	static int nbDoors; // Initialized to 0 and used to have successive id for the doors
-	int idDoor;
-};
-
-Door::nbDoors = 0;
