--- sources/main.cpp
@@ -1,5 +1,0 @@
-#include "Grid.h"
-
-int main(){
-    
-}
